// types.ts
export interface Comment {
  id: string;
  blogPostId: string;
  userId: string;
  userName: string; // Denormalized
  userProfilePictureUrl?: string; // Denormalized
  content: string;
  createdAt: string; // ISO Date string
  parentId?: string | null; // For replies
  replies?: Comment[]; // Nested replies
  reported?: boolean;
}
// ...
export interface AiContentSuggestion {
  suggestion: string;
}

export interface GeneratedImage {
  base64Image: string; // base64 encoded image data
  promptUsed: string;
}

// AI Generated Outline
export interface OutlineItem {
  type: 'h2' | 'h3' | 'h4' | 'point'; // 'point' for list items or paragraph points
  text: string;
  children?: OutlineItem[]; // For nested points or sub-subheadings
}

export interface AiFirstDraft {
    title: string;
    content: string; // HTML content
}